requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
Pillow==10.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
torch
torchvision
numpy
ultralytics
opencv-python==4.8.1.78
psycopg2-binary==2.9.9
